Minor fixes to the pluggable exchange, including pushing unix only parameters to the default implementation.
