+++ nbgrader/exchange/abc/exchange.py
@@ -19,17 +19,6 @@ class ExchangeError(Exception):
 
 
 class Exchange(LoggingConfigurable):
-
-    assignment_dir = Unicode(
-        ".",
-        help=dedent(
-            """
-            Local path for storing student assignments.  Defaults to '.'
-            which is normally Jupyter's notebook_dir.
-            """
-        )
-    ).tag(config=True)
-
     timezone = Unicode(
         "UTC",
         help="Timezone for recording timestamps"
@@ -38,32 +27,6 @@ class Exchange(LoggingConfigurable):
     timestamp_format = Unicode(
         "%Y-%m-%d %H:%M:%S.%f %Z",
         help="Format string for timestamps"
-    ).tag(config=True)
-
-    root = Unicode(
-        "/srv/nbgrader/exchange",
-        help="The nbgrader exchange directory writable to everyone. MUST be preexisting."
-    ).tag(config=True)
-
-    cache = Unicode(
-        "",
-        help="Local cache directory for nbgrader submit and nbgrader list. Defaults to $JUPYTER_DATA_DIR/nbgrader_cache"
-    ).tag(config=True)
-
-    @default("cache")
-    def _cache_default(self):
-        return os.path.join(jupyter_data_dir(), 'nbgrader_cache')
-
-    path_includes_course = Bool(
-        False,
-        help=dedent(
-            """
-            Whether the path for fetching/submitting  assignments should be
-            prefixed with the course name. If this is `False`, then the path
-            will be something like `./ps1`. If this is `True`, then the path
-            will be something like `./course123/ps1`.
-            """
-        )
     ).tag(config=True)
 
     coursedir = Instance(CourseDirectory, allow_none=True)
@@ -85,21 +48,6 @@ class Exchange(LoggingConfigurable):
             self.fail("Invalid timezone: {}".format(self.timezone))
         self.timestamp = datetime.datetime.now(tz).strftime(self.timestamp_format)
 
-    def set_perms(self, dest, fileperms, dirperms):
-        all_dirs = []
-        for dirname, _, filenames in os.walk(dest):
-            for filename in filenames:
-                os.chmod(os.path.join(dirname, filename), fileperms)
-            all_dirs.append(dirname)
-
-        for dirname in all_dirs[::-1]:
-            os.chmod(dirname, dirperms)
-
-    def ensure_root(self):
-        """See if the exchange directory exists and is writable, fail if not."""
-        if not check_directory(self.root, write=True, execute=True):
-            self.fail("Unwritable directory, please contact your instructor: {}".format(self.root))
-
     def init_src(self):
         """Compute and check the source paths for the transfer."""
         raise NotImplementedError
@@ -112,23 +60,9 @@ class Exchange(LoggingConfigurable):
         """Actually do the file transfer."""
         raise NotImplementedError
 
-    def do_copy(self, src, dest, log=None):
-        pass
-
     def start(self):
-        if sys.platform == 'win32':
-            self.fail("Sorry, the exchange is not available on Windows.")
-
-        if not self.coursedir.groupshared:
-            # This just makes sure that directory is o+rwx.  In group shared
-            # case, it is up to admins to ensure that instructors can write
-            # there.
-            self.ensure_root()
         self.set_timestamp()
 
         self.init_src()
         self.init_dest()
         self.copy_files()
-
-    def ensure_directory(self, path, mode):
-        pass
