+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
-;addopts = --strict-markers
+addopts = --strict-markers
 markers =
     nbextensions
